+++ demo-apps/next/src/demo-apps-backend-sdk.ts
@@ -1,0 +1,6 @@
+import { createClient, getServiceUrl } from 'demo-apps-backend-sdk';
+
+const { graphqlUrl } = getServiceUrl('local');
+export const demoAppsBackendSdk = createClient({
+  url: graphqlUrl || 'http://localhost:4000/graphql',
+});